Create Line,Drag,Mute
Added the Buzzer and Mute Function to main reel. The buzzer will turn on if the line length is increasing while the drag is above 0. The buzzer can be muted by pressing either of the pushbutton on the reel UI. the butter will not sound again until the drag is set to 0.